Refactor Line location
